requirements.txt
Django==1.9.9
django-bower==5.1.0
django-ckeditor==5.1.0
django-colorful==1.2
django-sortedm2m==1.3.2
six==1.10.0
